Immutable, singly-linked list implementation.
This code is very similar to scala.List, with a few key differences:

1. It does not expose any "unsafe" methods.
2. It is invariant, whereas scala.List is covariant.
3. It uses subtyping to differentiate non-emptiness.
4. It provides Cats type class instances.
5. It does not currently provide a fast (mutable) builder.

The types defined here are as follows:

 - Lst[A] represents a list of zero-or-more elements.
 - Nel[A] represents a list of one-or-more elements.
 - El[A]  represents an empty list (exactly zero elements).

(Every Lst[A] is either a Nel[A] or an El[A].)

While it does not (yet) provide every single Scala collection
method, it provides a decent number of them. Type class instances
are provided for both Lst and Nel (for example, Nel has a Comonad
instance, whereas Lst only has a CoflatMap instance).