added some benchmarks for list.
Made the list implementation much more reliant on the ListBuilder in
order to get huge speed wins. It's kinda gross, but it makes List much
much faster. Not sure if we'll keep this or not.

As a result, we might have lost some of the benefit of having the Nel
exposed as a type.