# This is a combination of 6 commits. # This is the 1st commit message: Setting version to 0.5.2-SNAPSHOT
# This is the commit message #2:

optimizations on Range

# This is the commit message #3:

Refactor Range and Diet to generate values lazily

many of the operations in Range and Diet relied on eagerly generating
exhaustinve lists of elements. This commit refactors those to evaluate
their contents lazily, which should have a dramatic reduction of the
amount of garbage generated.

# This is the commit message #4:

Typos and whitespaces at the end of lines.

# This is the commit message #5:

More typo fixes, replaced the top comment for 'Heap' with the docs version.

# This is the commit message #6:

Initial commit for disjoint-set.