Removed the deprecated Streaming class
Projects that depend on Streaming are advised to use fs2, monix or
iteratees instead.

The change also implements an inverse for `Discrete[A]` that
flips the `succ` and `pred` values of the source instance.
Also considered `reverse`, but I think `inverse` fits better
considering the type is expressed in terms of `succ` and `pred`
which are inversed.

The change also implements a foldRight for `Range[A]` as the
`Diet[A]` relied on the foldRight method in Streaming. The new
implementation call foldRight directly on Range.

fixes #122