Add a fast immutable BitSet implementation (#177)
* Add a fast immutable BitSet implementation

* Two commits for oscar/bitset (#178)

* Add a lot of comments to BitSet.

This includes both high-level documentation as well as some low-level
implementation notes.

* Test BitSet against negative integers.

When we wrote the BitSet tests we didn't think we supported negative values.
But it turns out we do, we just weren't testing with any of them. This commit
ensures we're testing across the entire range of Int values.

* Disable ScalaStyle.

* Fix some typos, clean up some prose.

* More typos.

* Fix three bugs:

  1. On 2.13 the set interface has changed, so we need a compat interface to
     split code across 2.12- adn 2.13+.

  2. Our hashCode method had a subtle bug involving Int.MinValue.

  3. The code for increasing ScalaCheck's MinSuccessfulTests didn't work.

This commit fixes all three. I tested (2) with 10K tests to get more confidence
in the fix (I could replicate the bug every time at that level).

* Add return types.

* Fix testing bug under JS.

* improve test coverage

* use next() instead of next since it side-effects

* Add xor and set-diff operations to BitSet. (#181)

* Add xor and set-diff operations to BitSet.

Also adds law-checking, and order/ordering instances.

* Add an intersects method for good measure.

* Respond to review comments.

Adds () to next calls and removes the unused method leafFor.

* Remove LogicalBitSet.

I was playing around with this but it isn't ready for primetime yet.

* Add two more laws and fix a typo.

* fix some bugs, improve some tests

* try to expand test coverage

* more test improvements