Comparing Diet
The same Diet might have many shapes depending on the insert/delete/merge/split operations executed on it (since it is not balanced). However, two Diets are equal when their ranges are equal (We dont need to iterate every single value on it, only its ranges)

- The worst case for Diet is when each node has only one value (range[x,x]) and each node is separated to each other in exact one value( nodes: [x,x] [x + 2, x + 2] [x + 4, x+4] ....

In the test I am building two dies like this in reverse order to make sure their shapes are different, however, the ranges must be the same.